#398 added withSchema() to SqlRepositoryBuilder to enable users to create and use tables in a defined Schema.
As part of the change, the static table and sequence names in the FixedSchemaFactory class have been made private and access to them is made available through static getters (which check if a schema was defined).

The JaversSchemaManager has been modified to check for the javers tables if a schema has been defined.  If no schema has been defined, this resposibility to handed over to PolyJDBC.