What ? 1. Adding @Order(0) to JaversSpringDataAuditableRepositoryAspect.
Why ?
This aspect contains advices that run for all save/delete operations in
the application but since there is no order present with it. It takes
the default order which is highest precedence as it is using after-
advices. So giving it order 0 would help if some other advice needed to
execute before this.
That's why explicitly giving it 0.